# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrustChain Public API",
    "description": "External TrustChain developer API. Authenticate with an organization API key (`Authorization: Bearer tc_live_***` or `tc_test_***`).",
    "version": "0.1.0",
    "contact": {
      "name": "TrustChain Developer Platform"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.trustchain.local/api/public/v1",
      "description": "Public API v1"
    }
  ],
...
  "security": [
    {
      "ApiKeyBearer": []
    }
  ],
  "paths": {
...
    "/signatures": {
      "post": {
        "tags": [
          "Signatures"
        ],
        "summary": "Create signature",
        "operationId": "createSignature",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "documentId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "certificateId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "algorithm": {
                    "type": "string"
                  },
                  "contentHash": {
                    "type": "string",
                    "nullable": true
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signature": {
                      "type": "object"
                    },
                    "generatedPrivateKeyPem": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/signatures/{id}": {
      "get": {
        "tags": [
          "Signatures"
        ],
        "summary": "Get signature",
        "operationId": "getSignature",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signature",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signature": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Organization API key (`tc_live_…` / `tc_test_…`) or service-account secret (`sa_sec_…`)."
      }
    },
...
  }
}